requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
-e git+https://github.com/JuanPotato/Legofy#egg=Legofy
click==6.2                # via pip-tools
first==2.0.1              # via pip-tools
flask==0.10.1
gunicorn==19.4.1
itsdangerous==0.24        # via flask
jinja2==2.8               # via flask
markupsafe==0.23          # via jinja2
pillow==3.0.0
pip-tools==1.2.0
six==1.10.0               # via pip-tools
werkzeug==0.11.2          # via flask
